Add query in repository

+++ stat-service/dto/src/main/java/ru/practicum/ewm/ViewStats.java
@@ -12,5 +12,5 @@ import lombok.ToString;
 public class ViewStats {
     private String app;
     private String uri;
-    private int hits;
+    private long hits;
 }
